requirements.txt
Django==1.7
django-classy-tags==0.6.1
django-extensions==1.5.5
django-import-export==0.2.7
django-orderable==3.1.0
django-polymorphic==0.7.1
django-shop==0.2.0
jsonfield==1.0.3
psycopg2==2.6
pydot==1.0.2
pyparsing==1.5.7
six==1.9.0
South==1.0.2
tablib==0.10.0
